requirements.txt
Flask==2.3.3
Flask-RESTful==0.3.9
PyPDF2==3.0.1
python-docx==0.8.11
networkx==3.1
requests==2.31.0
python-dotenv==1.0.0